passwordless login

+++ src/server.py
@@ -4,6 +4,7 @@ import json
 from datetime import datetime, timedelta
 import log
 import uuid
+import urllib
 
 logger = log.setup('root', 'sxsw.log')
 from email.utils import parseaddr
@@ -19,7 +20,6 @@ static_root = "sxsw"
 protocol = "http://"
 base_url = protocol + domain + "/" + app_root
 homepage = protocol + domain + "/" + static_root + "/"
-loginpage = homepage + "login.html"
 cookiedomain = domain
 port = 8081
 
@@ -135,15 +135,29 @@ def test_db():
 
 test_db()
 
+def check_user():
+    #check is user is logged in
+    auth_cookie = request.get_cookie('AUTH_COOKIE')
+    logger.debug(auth_cookie)
+    if auth_cookie and os.path.isfile(authdir+auth_cookie):
+        logger.debug("file exists")
+        c.execute("SELECT * FROM sxsw_users WHERE uuid=?", (auth_cookie,))
+        user = c.fetchone()
+        return user
+    else:
+        return
+
 
 @post('/users/<userid>/events/<eventid>')
 def index(userid, eventid):
-    logger.debug(eventid)
-    c.execute("REPLACE INTO sxsw_user_events VALUES (?,?)", (userid, eventid))
-    conn.commit()
-    if update_elastic(eventid):
-        return "ACK"
-    return "NACK"
+    user = check_user()
+    if user:
+        logger.debug(eventid)
+        c.execute("REPLACE INTO sxsw_user_events VALUES (?,?)", (userid, eventid))
+        conn.commit()
+        if update_elastic(eventid):
+            return "ACK"
+    return "Not logged in"
 
 
 def update_elastic(eventid):
@@ -173,11 +187,13 @@ def update_elastic(eventid):
 @delete('/users/<userid>/events/<eventid>')
 def index(userid, eventid):
     logger.debug(eventid)
-    c.execute("DELETE FROM sxsw_user_events WHERE userid=? AND eventid=?", (userid, eventid))
-    conn.commit()
-    if update_elastic(eventid):
-        return "ACK"
-    return "NACK"
+    user = check_user()
+    if user:
+        c.execute("DELETE FROM sxsw_user_events WHERE userid=? AND eventid=?", (userid, eventid))
+        conn.commit()
+        if update_elastic(eventid):
+            return "ACK"
+    return "Not logged in"
 
 
 def get_all_attendees(eventid):
@@ -207,10 +223,10 @@ def index(userid):
     return json.dumps(r)
 
 
-def check_admin_rights(u_id):
-    res = c.fetchone("SELECT * FROM users WHERE uuid=%s AND admin=1", (u_id,))
-    return res
-
+def get_all_users():
+    c.execute("SELECT handle FROM sxsw_users ORDER BY handle")
+    users = c.fetchall()
+    return users
 
 @get("/api/auth/check", apply=[jsonp_plugin])
 def auth_check():
@@ -229,18 +245,12 @@ def auth_check():
         logger.debug(res)
         if res:
             logger.debug("check ok")
-#            if not (user):
-#                response.delete_cookie("AUTH_COOKIE")
-#                result = {"loggedin": False, "loginpage": loginpage}
-#                logger.debug(result)
-#                return result
             response.set_cookie("AUTH_COOKIE", auth_cookie, expires=expires, domain=cookiedomain, path="/")
             result = {"user_token": res.get("user_token"),
                       "csrf_token": res.get("csrf_token"), "loggedin": True, "email": res["email"],"userid":res["userid"],"handle":res["handle"]}
         else:
-            #        start_response("200 OK",[("Content-type", "text/html"),("Set-Cookie","AUTH_COOKIE=no_access; Domain=cl.gbme.nl; Path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT")])
             response.delete_cookie("AUTH_COOKIE")
-            result = {"loggedin": False, "loginpage": "/sxsw/login.html"}
+            result = {"loggedin": False}
             logger.debug(result)
             logger.debug(result)
         return result
@@ -262,7 +272,7 @@ def input_cleaner(text):
     return text.replace("<", "")
 
 
-def send_simple_message(email, token, template, handler):
+def send_simple_message(email, token, template, handler,subject):
     template = open("/vagrant/src/templates/" + template).read()
     if template:
         text = template.replace("{{url}}", homepage + "/" + handler + "?email=" + email + "&token=" + token)
@@ -271,10 +281,22 @@ def send_simple_message(email, token, template, handler):
             auth=("api", "key-9fa82127b18561758f4809985b438922"),
             data={"from": "gbme <gerbrand@gbme.nl>",
                   "to": [email],
-                  "subject": "sxsw browser invite",
+                  "subject": subject,
                   "html": text})
     else:
         throw("no template found")
+
+
+def logout(message = False):
+    auth_cookie = request.get_cookie('AUTH_COOKIE')
+    logger.debug("logout")
+    if auth_cookie:
+        logger.debug("remove cookie:"+CONFIG.global_config.cookie_store+auth_cookie)
+        rm(CONFIG.global_config.cookie_store+auth_cookie)
+    response.status = 200
+    response.add_header("Content-type", "application/json")
+    response.delete_cookie("AUTH_COOKIE")
+    return json.dumps({"result": "success", "login_status": False,"message":message})
 
 
 @post("/api/auth/new_user/", apply=[jsonp_plugin])
@@ -283,80 +305,158 @@ def auth_new_user():
         data = request.json
         logger.debug(data)
         invite_token = data.get("token")
-        email = data.get("userid").replace("%40", "@")
-        passwd = data.get("new_password")
+        email= urllib.unquote(data.get("userid" )).decode('utf8')
         handle = data.get("handle")
-        hash = pwd_context.encrypt(passwd)
-        logger.debug(hash)
         logger.debug(email)
         logger.debug(invite_token)
+        logger.debug(handle)
         res = c.execute("SELECT user_token FROM sxsw_users WHERE invite_token=? AND email=?", (invite_token, email))
         res = c.fetchone()
         logger.debug(res)
         if res:
-            expires = (datetime.today() + timedelta(minutes=3600)).strftime('%H:%M:%S-%a/%d/%b/%Y')
             my_uuid = unique_id()
             logger.debug("uuid:" + my_uuid)
             c.execute(
-                "UPDATE sxsw_users SET hash=?,invite_token=NULL,invite_created=0,active=1,handle=? WHERE hash IS NULL AND invite_token=? AND email=?",
-                (hash, handle, invite_token, email))
+                "UPDATE sxsw_users SET invite_token=NULL,invite_created=0,active=1,handle=? WHERE hash IS NULL AND invite_token=? AND email=?",
+                (handle, invite_token, email))
             conn.commit()
             c.execute("SELECT * FROM sxsw_users WHERE email=?",(email,))
             logger.debug(c.fetchone())
             result = set_login_cookie(my_uuid, res["user_token"])
-            c.execute("SELECT * FROM sxsw_users WHERE email=?",(email,))
-            logger.debug(c.fetchone())
             return result
 
+        else:
+            return logout("token expired?")
+    except:
+        logger.exception("er")\
+
+#
+# @post("/xapi/auth/new_user/", apply=[jsonp_plugin])
+# def auth_new_user():
+#     try:
+#         data = request.json
+#         logger.debug(data)
+#         invite_token = data.get("token")
+#         email= urllib.unquote(data.get("userid" )).decode('utf8')
+#         passwd = data.get("new_password")
+#         handle = data.get("handle")
+#         hash = pwd_context.encrypt(passwd)
+#         logger.debug(hash)
+#         logger.debug(email)
+#         logger.debug(invite_token)
+#         res = c.execute("SELECT user_token FROM sxsw_users WHERE invite_token=? AND email=?", (invite_token, email))
+#         res = c.fetchone()
+#         logger.debug(res)
+#         if res:
+#             expires = (datetime.today() + timedelta(minutes=3600)).strftime('%H:%M:%S-%a/%d/%b/%Y')
+#             my_uuid = unique_id()
+#             logger.debug("uuid:" + my_uuid)
+#             c.execute(
+#                 "UPDATE sxsw_users SET hash=?,invite_token=NULL,invite_created=0,active=1,handle=? WHERE hash IS NULL AND invite_token=? AND email=?",
+#                 (hash, handle, invite_token, email))
+#             conn.commit()
+#             c.execute("SELECT * FROM sxsw_users WHERE email=?",(email,))
+#             logger.debug(c.fetchone())
+#             result = set_login_cookie(my_uuid, res["user_token"])
+#             return result
+#
+#         else:
+#             return logout()
+#     except:
+#         logger.exception("er")
+
+
+
+
+# @post("/xapi/auth/login/", apply=[jsonp_plugin])
+# def auth_login():
+#     try:
+#         auth_cookie = request.get_cookie('AUTH_COOKIE')
+#         logger.debug(auth_cookie)
+#         data = request.json
+#         if data and "userid" in data and "password" in data:
+#             userid = data["userid"]
+#             passwd = data["password"]
+#             logger.debug(userid)
+#             c.execute("SELECT * FROM sxsw_users WHERE email=? AND active=1", (userid,))
+#             res = c.fetchone()
+#             logger.debug(res)
+#             if res and res["hash"]:
+#                 logger.debug("user exists")
+#                 testpw = pwd_context.verify(passwd, res["hash"])
+#                 if testpw:
+#                     if res["uuid"] and len(res["uuid"]) > 0:
+#                         my_uuid = res["uuid"]
+#                         pass
+#                     else:
+#                         my_uuid = unique_id()
+#                         logger.debug("uuid:" + my_uuid)
+#                     return set_login_cookie(my_uuid, res["user_token"])
+#                 else:
+#                     logger.debug("wrong pw")
+#         logger.error("not a valid login request")
+#         # else remove cookie
+#         if auth_cookie:
+#             rm(authdir + auth_cookie)
+#         logger.debug("error in auth, logout")
+#         response.status = 200
+#         response.add_header("Content-type", "application/json")
+#         response.delete_cookie("AUTH_COOKIE")
+#         return json.dumps(
+#             {"result": "failed", "login_status": False, "Location": loginpage + "?error=wrong_credentials"})
+#     except:
+#         logger.exception("error")
+#         response.status = 501
+#         return "Internal server error"
+
+@post("/api/auth/reset_password/")
+def auth_resetpw():
+    try:
+        data = request.json
+        logger.debug(data)
+        token = data.get("token" )
+        email= urllib.unquote(data.get("userid" )).decode('utf8')
+        logger.debug(email)
+        logger.debug(token)
+        c.execute("SELECT user_token FROM sxsw_users WHERE reset_token=? AND email=? ",(token,email))
+        logger.debug("SELECT user_token FROM sxsw_users WHERE reset_token=%s AND email=%s AND reset_date > date('now','10 minutes')",(token,email))
+        res = c.fetchone()
+        logger.debug(res)
+        if res:
+            my_uuid = unique_id()
+            logger.debug("uuid:"+my_uuid)
+            c.execute("UPDATE sxsw_users SET reset_token=NULL,reset_date=NULL,active=1 WHERE reset_token=? AND email=?",(token,email))
+            conn.commit()
+            return set_login_cookie(my_uuid,res["user_token"])
+            return
         else:
             return logout()
     except:
         logger.exception("er")
 
 
-@post("/api/auth/login/", apply=[jsonp_plugin])
+@post("/api/auth/logout/", apply=[jsonp_plugin])
 def auth_login():
     try:
         auth_cookie = request.get_cookie('AUTH_COOKIE')
-        logger.debug(auth_cookie)
-        data = request.json
-        if data and "userid" in data and "password" in data:
-            userid = data["userid"]
-            passwd = data["password"]
-            logger.debug(userid)
-            c.execute("SELECT * FROM sxsw_users WHERE email=? AND active=1", (userid,))
-            res = c.fetchone()
-            logger.debug(res)
-            if res and res["hash"]:
-                logger.debug("user exists")
-                testpw = pwd_context.verify(passwd, res["hash"])
-                if testpw:
-                    my_uuid = None
-
-                    if res["uuid"] and len(res["uuid"]) > 0:
-                        my_uuid = res["uuid"]
-                        pass
-                    else:
-                        my_uuid = unique_id()
-                        logger.debug("uuid:" + my_uuid)
-                    return set_login_cookie(my_uuid, res["user_token"])
-                else:
-                    logger.debug("wrong pw")
-        logger.error("not a valid login request")
-        # else remove cookie
         if auth_cookie:
             rm(authdir + auth_cookie)
         logger.debug("error in auth, logout")
-        #        start_response("301 Moved Permanently",[("Location","http://bb.gbme.nl/static/ajax/login.html?error=wrong_credentials"),("Content-type", "text/html"),("Set-Cookie","AUTH_COOKIE=no_access; Domain=.gbme.nl; Path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT")])
         response.status = 200
         response.add_header("Content-type", "application/json")
         response.delete_cookie("AUTH_COOKIE")
         return json.dumps(
-            {"result": "failed", "login_status": False, "Location": loginpage + "?error=wrong_credentials"})
+            {"result": "ok", "login_status": False})
     except:
         logger.exception("error")
         response.status = 501
         return "Internal server error"
+
+def rm(fname, times=None):
+    try:
+        os.remove(fname)
+    except:
+        logger.debug("file "+fname+ " not found")
 
 
 def set_login_cookie(my_uuid, user_token):
@@ -375,37 +475,75 @@ def set_login_cookie(my_uuid, user_token):
     return {"result": "ok", "login_status": True, "Location": homepage}
 
 
-@post("/api/v1/admin/users")
+@post("/api/users/invite/")
 def admin_users():
-    logger.debug("test")
-    # user = check_admin_rights(request.get_cookie("AUTH_COOKIE"))
-    user = 1
-    if not (user):
-        return
-    data = request.json
-    logger.debug(data)
-    email = test_email(data["new_email"])
-    if not (email):
-        return {"result": "invalid email", "message": "Ongeldig email adres"}
+    user = check_user()
+    if user:
+        logger.debug("test")
+        # user = check_admin_rights(request.get_cookie("AUTH_COOKIE"))
+        data = request.json
+        logger.debug(data)
+        email = test_email(data["new_email"])
+        if not (email):
+            return {"result": "invalid email", "message": "Ongeldig email adres"}
 
-    invite_token = unique_id()
-    user_token = unique_id()
-    csrf_token = unique_id()
-    # check email
+        invite_token = unique_id()
+        user_token = unique_id()
+        csrf_token = unique_id()
+        # check email
+        try:
+            print (email, user_token, invite_token, csrf_token)
+            new_id = c.execute(
+                "INSERT INTO sxsw_users (email,user_token,invite_token,csrf_token,invite_created) VALUES (?,?,?,?,date('now'))",
+                (email, user_token, invite_token, csrf_token))
+            conn.commit()
+            send_simple_message(email, invite_token, "mail_invite", "new_user.html","SXSW browser invite")
+            #     res = user_list()
+            return {"result": "ok", "data": ""}
+        except:
+            logger.exception("duplicate user")
+            return {"result": "duplicate user", "message": "Er is al een gebruiker met dit email adres"}
+    return {"result":"failed","message":"not logged in"}
+
+@post("/api/auth/forgot/")
+def auth_forgot():
     try:
-        print (email, user_token, invite_token, csrf_token)
-        new_id = c.execute(
-            "INSERT INTO sxsw_users (email,user_token,invite_token,csrf_token,invite_created) VALUES (?,?,?,?,date('now'))",
-            (email, user_token, invite_token, csrf_token))
-        conn.commit()
-        send_simple_message(email, invite_token, "mail_invite", "new_user.html")
-        #     res = user_list()
-        return {"result": "ok", "data": ""}
+        auth_cookie = request.get_cookie('AUTH_COOKIE')
+        if auth_cookie:
+            logger.debug("logout user:"+auth_cookie)
+            rm(authdir+auth_cookie)
+        logger.debug(auth_cookie)
+
+        data = request.json
+        logger.debug(data)
+        userid = data.get("userid")
+        if userid:
+            logger.debug("userid:"+str(userid))
+            c.execute("SELECT * from sxsw_users WHERE email=? AND active=1",(userid,))
+            res = c.fetchone()
+            logger.debug(res)
+            if res:
+                logger.debug("user exists")
+                reset_token = str(unique_id())
+                send_simple_message(res["email"],reset_token,"mail_resetpw","reset_password.html","SXSW browser login link")
+                #check email
+                c.execute("UPDATE sxsw_users SET reset_token=?,reset_date=date('now') WHERE email=?",(reset_token,res["email"]))
+                conn.commit()
+                c.execute("SELECT * FROM sxsw_users WHERE email=?",(res["email"],))
+                logger.debug(c.fetchone())
+                return {"result":True,"message":"Check your mailbox!"}
+    #        start_response("301 Moved Permanently",[("Location","http://bb.gbme.nl/static/ajax/login.html?error=wrong_credentials"),("Content-type", "text/html"),("Set-Cookie","AUTH_COOKIE=no_access; Domain=.gbme.nl; Path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT")])
+            response.delete_cookie("AUTH_COOKIE")
+            logger.debug("no such user:"+userid)
+            return {"result":True,"message":"check mailbox"}
     except:
-        logger.exception("duplicate user")
-        return {"result": "duplicate user", "message": "Er is al een gebruiker met dit email adres"}
+        logger.exception("error")
+        response.status = 501
+        return "Internal server error"
 
 
 logger.debug("starting server at:" + str(port))
 
 run(host='localhost', port=port)
+
+
